I've implemented logout only when theres a internet connectivity (feature). But please note that its gonna work kind of real-time, i mean the isConnected is being collected in real time which means if internet status changes current job's destroy and new one according state launched.

+++ composeApp/src/commonMain/kotlin/me/androidbox/settings/presentation/SettingsViewModel.kt
@@ -4,22 +4,28 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import me.androidbox.ConnectivityManager
 import me.androidbox.NoteMarkPreferences
 import me.androidbox.authentication.login.domain.model.LogoutRequest
 import me.androidbox.authentication.login.domain.use_case.LogoutUseCase
 import me.androidbox.notes.domain.usecases.NukeAllNotesUseCase
-import me.androidbox.settings.presentation.SettingsEvent.*
+import me.androidbox.settings.presentation.SettingsEvent.logoutSuccess
+import me.androidbox.settings.presentation.SettingsEvent.onShowMessage
 import net.orandja.either.Left
 import net.orandja.either.Right
 
 class SettingsViewModel(
     private val logoutUseCase: LogoutUseCase,
     private val nukeAllNotesUseCase: NukeAllNotesUseCase,
-    private val noteMarkPreferences: NoteMarkPreferences
+    private val noteMarkPreferences: NoteMarkPreferences,
+    private val connectivityManager: ConnectivityManager
 ) : ViewModel() {
     private val _state = MutableStateFlow(SettingsScreenUiState())
     val state = _state.asStateFlow()
@@ -27,29 +33,47 @@ class SettingsViewModel(
     private val _settingsEvent = Channel<SettingsEvent>()
     val settingsEvent = _settingsEvent.receiveAsFlow()
 
+    private val isConnected = connectivityManager
+        .isConnected()
+        .stateIn(
+            viewModelScope,
+            SharingStarted.WhileSubscribed(5000),
+            false
+        )
+
     fun action(settingsAction: SettingsAction) {
         when (settingsAction) {
             SettingsAction.OnLogout -> {
                 viewModelScope.launch {
-                    val refreshToken = noteMarkPreferences.getRefreshToken()
+                    isConnected.collectLatest { connected ->
+                        if (connected) {
+                            val refreshToken = noteMarkPreferences.getRefreshToken()
 
-                    if (refreshToken != null) {
-                        val result = logoutUseCase.execute(
-                            logoutRequest = LogoutRequest(
-                                refreshToken = refreshToken
+                            if (refreshToken != null) {
+                                val result = logoutUseCase.execute(
+                                    logoutRequest = LogoutRequest(
+                                        refreshToken = refreshToken
+                                    )
+                                )
+
+                                when (result) {
+                                    is Left -> {
+                                        noteMarkPreferences.deleteAllPreferences()
+                                        nukeAllNotesUseCase.execute()
+                                        _settingsEvent.send(logoutSuccess(isSuccess = true))
+                                    }
+
+                                    is Right -> {
+                                        _settingsEvent.send(logoutSuccess(isSuccess = false))
+                                    }
+                                }
+                            }
+                        } else {
+                            _settingsEvent.trySend(
+                                onShowMessage(
+                                    message = "You need an internet connection to log out."
+                                )
                             )
-                        )
-
-                        when (result) {
-                            is Left -> {
-                                noteMarkPreferences.deleteAllPreferences()
-                                nukeAllNotesUseCase.execute()
-                                _settingsEvent.send(logoutSuccess(isSuccess = true))
-                            }
-
-                            is Right -> {
-                                _settingsEvent.send(logoutSuccess(isSuccess = false))
-                            }
                         }
                     }
                 }
